ci: fix vlib/term/ui/ui_test.v compilation

+++ vlib/term/ui/termios_nix.c.v
@@ -12,9 +12,9 @@ import time
 
 struct C.termios {
 mut:
-	c_iflag u32
-	c_lflag u32
-	c_cc    [32]byte
+	c_iflag int
+	c_lflag int
+	c_cc    [32]int
 }
 
 struct C.winsize {
@@ -74,11 +74,11 @@ fn (mut ctx Context) termios_setup() ? {
 	if ctx.cfg.capture_events {
 		// Set raw input mode by unsetting ICANON and ECHO,
 		// as well as disable e.g. ctrl+c and ctrl.z
-		termios.c_iflag &= ~u32(C.IGNBRK | C.BRKINT | C.PARMRK | C.IXON)
-		termios.c_lflag &= ~u32(C.ICANON | C.ISIG | C.ECHO | C.IEXTEN | C.TOSTOP)
+		termios.c_iflag &= ~(C.IGNBRK | C.BRKINT | C.PARMRK | C.IXON)
+		termios.c_lflag &= ~(C.ICANON | C.ISIG | C.ECHO | C.IEXTEN | C.TOSTOP)
 	} else {
 		// Set raw input mode by unsetting ICANON and ECHO
-		termios.c_lflag &= ~u32(C.ICANON | C.ECHO)
+		termios.c_lflag &= ~(C.ICANON | C.ECHO)
 	}
 
 	if ctx.cfg.hide_cursor {
